requirements.txt
numpy
matplotlib
opencv-python
pillow

